fix ADQL build file to permit cross-compilation
You can now set the adql.bootclasspath system property to use
an alternative bootclasspath, for instance that for a different
JRE than that used by the compiler.

For instance, this cross-compiles to java 1.5:

   ant -Dant.build.javac.target=1.5
       -Dadql.bootclasspath=/usr/java/jdk1.5.0_22/jre/lib/rt.jar
       -f buildADQL.xml
       buildLib

Note the java used by ant has to be at least java 6 because of the
@Override source constructs on methods inherited from interfaces
(only methods from abstract or non-abstract classes can be annotated
with @Override in java 5).